LPS-195551 Move interfaces to asset-display-page-* since it makes more sense to have them there. Change namings and correct package
