+++ modules/apps/asset/asset-display-page-api/src/main/java/com/liferay/asset/display/page/layout/asset/entry/provider/LayoutAssetEntryProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.provider;
+package com.liferay.asset.display.page.layout.asset.entry.provider;
 
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.portal.kernel.model.Layout;
@@ -13,7 +13,7 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Roberto Díaz
  */
-public interface PortletLayoutAssetEntryProvider {
+public interface LayoutAssetEntryProvider {
 
 	public AssetEntry getLayoutAssetEntry(
 		HttpServletRequest httpServletRequest, Layout layout);
