+++ modules/apps/asset/asset-display-page-service/src/main/java/com/liferay/asset/display/page/internal/layout/asset/entry/provider/LayoutAssetEntryProviderRegistryImpl.java
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.internal.util;
+package com.liferay.asset.display.page.internal.layout.asset.entry.provider;
 
-import com.liferay.asset.provider.PortletLayoutAssetEntryProvider;
-import com.liferay.asset.provider.PortletLayoutAssetEntryProviderRegistry;
+import com.liferay.asset.display.page.layout.asset.entry.provider.LayoutAssetEntryProvider;
+import com.liferay.asset.display.page.layout.asset.entry.provider.LayoutAssetEntryProviderRegistry;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 
@@ -19,22 +19,22 @@ import org.osgi.service.component.annotations.Modified;
 /**
  * @author Roberto Díaz
  */
-@Component(service = PortletLayoutAssetEntryProviderRegistry.class)
-public class PortletLayoutAssetEntryProviderRegistryImpl
-	implements PortletLayoutAssetEntryProviderRegistry {
+@Component(service = LayoutAssetEntryProviderRegistry.class)
+public class LayoutAssetEntryProviderRegistryImpl
+	implements LayoutAssetEntryProviderRegistry {
 
 	@Override
-	public PortletLayoutAssetEntryProvider getPortletLayoutAssetEntryProvider(
-		String className) {
+	public LayoutAssetEntryProvider getLayoutAssetEntryProvider(
+		String portletId) {
 
-		return _serviceTrackerMap.getService(className);
+		return _serviceTrackerMap.getService(portletId);
 	}
 
 	@Activate
 	@Modified
 	protected void activate(BundleContext bundleContext) {
 		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, PortletLayoutAssetEntryProvider.class,
+			bundleContext, LayoutAssetEntryProvider.class,
 			"javax.portlet.name");
 	}
 
@@ -43,7 +43,7 @@ public class PortletLayoutAssetEntryProviderRegistryImpl
 		_serviceTrackerMap.close();
 	}
 
-	private volatile ServiceTrackerMap<String, PortletLayoutAssetEntryProvider>
+	private volatile ServiceTrackerMap<String, LayoutAssetEntryProvider>
 		_serviceTrackerMap;
 
 }
